+++ crucible/CHANGELOG.md
@@ -24,6 +24,8 @@
   traces in `SimError` and `IsSymBackend`
 * Fix `onlineProve` to properly negate goals before checking satisfiability
 * Avoid sending trivially-true goals to the solver in `Backend.Prove`.
+* Fix bug in `execResultGlobals` that caused it to erroneously always return the
+  non-aborted branch of a `FinishedResult ... (PartialRes ...)`.
 
 # 0.9 -- 2026-01-29